Draft commit. I'm trying to reorganize this code, so that we fix all the network issues. It's far from compiling, very far from running fine and very very far from being stable.
