+++ tc65lib/src/org/javacint/control/m2mp/M2MPClient.java
@@ -1,26 +1,16 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package org.javacint.control.m2mp;
 
-/**
- *
- * @author Florent
- */
-public interface M2MPClient extends IProtocolLayerSend {
+public interface M2MPClient extends M2MPSender {
 
-	void setListener(IProtocolLayerReceive listener);
+    void setListener(M2MPEventsListener listener);
 
-	void start() throws Exception;
+    void start() throws Exception;
 
-	void stop() throws Exception;
+    void stop() throws Exception;
 
-	public void setIdent(String string);
+    public void setIdent(String string);
 
-	public void setCapabilities(String echoloc);
+    public void setCapabilities(String echoloc);
 
-	public void sendCapabilities();
-
-	public void setStatus(String name, String value);
+    public void setStatus(String name, String value);
 }
