+++ tc65lib/src/org/javacint/control/m2mp/README
@@ -1,0 +1,5 @@
+This is the M2MP implementation.
+
+It made some bad design choices and a quite bad implementation. I'm trying to
+do a rewrite of this codebase because it's used on the field and its concept
+are still valid.
